+++ psycop/common/model_training/config_schemas/data.py
@@ -34,8 +34,12 @@ class DataSchema(PSYCOPBaseModel):
         "train",
         "val",
     ]  # splits to use for training
-    splits_for_evaluation: Optional[Sequence[Literal["val", "test"]]] = None
-    # splits to use for evaluation. If None, crossvalidation is done on splits_for_training split(s).
+    datasets_for_evaluation: Optional[
+        Sequence[Union[Literal["val", "test"], str]]
+    ] = None
+    # dataset to use for evaluation. Commonly, it would simply take either the val or test split
+    # of the corresponding train split, but a specific dataset may also be given when needed.
+    # If None, crossvalidation is done on splits_for_training split(s).
 
     # Feature specs
     col_name: ColumnNamesSchema = ColumnNamesSchema()